Wraps standard statistical operations by a slightly more elaborate structure.
Standard statistical operations are defined in 'std-ops.rkt' as normal
procedures. In main.rkt, we wrap these normal procedures with an
abstraction called stat-operation. This to imply certain constraints of
statistical operations. For instance, some operations require that the
datasets must be sorted. Yet, we don't want to sort the dataset each time
we call an operation on an arbitrary dataset.

Statistical datasets are also wrapped with an abstraction called
stat-dataset. This abstraction creates a cache for already-applied
operations. The current limitation is that, in order for the cache
to be used, it should be used with augmented operations and datasets;
i.e. instances of stat-operation and stat-dataset, respectively.

Also adds .giignore.